requirements.txt
Django
dj-database-url
django-compressor
django-extensions
django-haystack
django-imagekit
django-libsass
django-recaptcha
lxml
Markdown
pysolr
requests
psycopg2
webob
oauth2client